+++ bevy_asset_loader/examples/full_collection.rs
@@ -1,6 +1,7 @@
 use bevy::app::AppExit;
 use bevy::asset::LoadState;
 use bevy::prelude::*;
+use bevy::utils::HashMap;
 use bevy_asset_loader::prelude::*;
 
 fn main() {
@@ -40,12 +41,27 @@ struct MyAssets {
     // A folder loaded to typed asset handles (not supported on the web)
     #[asset(path = "images", collection(typed))]
     folder_typed: Vec<Handle<Image>>,
+    // A folder loaded as map (not supported on the web)
+    #[asset(path = "images", collection(mapped))]
+    mapped_folder_untyped: HashMap<String, HandleUntyped>,
+    // A folder loaded to typed asset handles mapped with their file names (not supported on the web)
+    #[asset(path = "images", collection(typed, mapped))]
+    mapped_folder_typed: HashMap<String, Handle<Image>>,
     // A collection of asset files
     #[asset(paths("images/player.png", "images/tree.png"), collection)]
     files_untyped: Vec<HandleUntyped>,
     // A collection of asset files loaded to typed asset handles
     #[asset(paths("images/player.png", "images/tree.png"), collection(typed))]
     files_typed: Vec<Handle<Image>>,
+    // A mapped collection of asset files
+    #[asset(paths("images/player.png", "images/tree.png"), collection(mapped))]
+    mapped_files_untyped: HashMap<String, HandleUntyped>,
+    // A mapped collection of asset files loaded to typed asset handles
+    #[asset(
+        paths("images/player.png", "images/tree.png"),
+        collection(typed, mapped)
+    )]
+    mapped_files_typed: HashMap<String, Handle<Image>>,
 }
 
 fn expectations(
@@ -98,6 +114,38 @@ fn expectations(
             LoadState::Loaded
         );
     }
+    assert_eq!(assets.mapped_folder_untyped.len(), 6);
+    for (name, handle) in assets.mapped_folder_untyped.iter() {
+        assert_eq!(
+            asset_server.get_load_state(handle.clone()),
+            LoadState::Loaded
+        );
+        assert_eq!(
+            asset_server
+                .get_handle_path(handle.clone())
+                .unwrap()
+                .path()
+                .to_str()
+                .unwrap(),
+            &format!("images/{name}")
+        );
+    }
+    assert_eq!(assets.mapped_folder_typed.len(), 6);
+    for (name, handle) in assets.mapped_folder_typed.iter() {
+        assert_eq!(
+            asset_server.get_load_state(handle.clone()),
+            LoadState::Loaded
+        );
+        assert_eq!(
+            asset_server
+                .get_handle_path(handle.clone())
+                .unwrap()
+                .path()
+                .to_str()
+                .unwrap(),
+            &format!("images/{name}")
+        );
+    }
     assert_eq!(assets.files_untyped.len(), 2);
     for handle in assets.files_untyped.iter() {
         assert_eq!(
@@ -110,6 +158,38 @@ fn expectations(
         assert_eq!(
             asset_server.get_load_state(handle.clone()),
             LoadState::Loaded
+        );
+    }
+    assert_eq!(assets.mapped_files_untyped.len(), 2);
+    for (name, handle) in assets.mapped_files_untyped.iter() {
+        assert_eq!(
+            asset_server.get_load_state(handle.clone()),
+            LoadState::Loaded
+        );
+        assert_eq!(
+            name,
+            asset_server
+                .get_handle_path(handle.clone())
+                .unwrap()
+                .path()
+                .to_str()
+                .unwrap()
+        );
+    }
+    assert_eq!(assets.mapped_files_typed.len(), 2);
+    for (name, handle) in assets.mapped_files_typed.iter() {
+        assert_eq!(
+            asset_server.get_load_state(handle.clone()),
+            LoadState::Loaded
+        );
+        assert_eq!(
+            name,
+            asset_server
+                .get_handle_path(handle.clone())
+                .unwrap()
+                .path()
+                .to_str()
+                .unwrap()
         );
     }